fix: Mount extra volumes as rw (needed for cache writes), fix docs and types
- Revert extra volumes to read-write mode (app needs to write cache files)
- Update docstrings and config descriptions to say read-write, not read-only
- Cast setup_timeout to int to match exec_command signature

+++ src/mcpbr/harnesses.py
@@ -900,7 +900,7 @@ class ClaudeCodeHarness:
         # pre-computing caches that should not count against timeout_seconds.
         if self.mcp_server and self.mcp_server.setup_command:
             setup_cmd = self.mcp_server.get_setup_command_for_workdir(env.workdir)
-            setup_timeout = self.mcp_server.setup_timeout_ms / 1000
+            setup_timeout = int(self.mcp_server.setup_timeout_ms / 1000)
 
             if verbose:
                 self._console.print(
